requirements.txt
vllm==0.16.0
runpod==1.6.2
huggingface-hub
pydantic
pillow
requests
pyyaml
